% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.negativeVirome.R
\name{get.negativeVirome}
\alias{get.negativeVirome}
\title{A wrapper of get* functions to create an empty palm.virome data.frame
which includes all runs analyzed in Serratus, not simply those with viruses
detected. Used to calculate how manmy "Virus Negative" datasets are in
a search.}
\usage{
get.negativeVirome(run.vec = NA, org.search = NA, con = SerratusConnect())
}
\arguments{
\item{run.vec}{character vector, sra run accession list}

\item{org.search}{string, search term to query "scientific_name". Uses SQL "like"}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
palm.virome  data.frame
}
\description{
A wrapper of get* functions to create an empty palm.virome data.frame
which includes all runs analyzed in Serratus, not simply those with viruses
detected. Used to calculate how manmy "Virus Negative" datasets are in
a search.
}
\examples{

con <- SerratusConnect()

\donttest{
get.negativeVirome(org.search = 'Drosophilia\%', con )
}

}
\keyword{Serratus}
\keyword{bioproject}
\keyword{biosample}
\keyword{control}
\keyword{geo}
\keyword{palmid}
\keyword{sql}
\keyword{sra}
\keyword{timeline}
