% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sraMeta.R
\name{get.sraMeta}
\alias{get.sraMeta}
\title{get.sraMeta}
\usage{
get.sraMeta(run_ids, con, ordinal = FALSE)
}
\arguments{
\item{run_ids}{character, SRA 'run_id'}

\item{con}{pq-connection, use SerratusConnect()}

\item{ordinal}{boolean, return 'run_ids' ordered vector [FALSE]}
}
\value{
data.frame, SRA Run Metadata
}
\description{
Retrieve the "BioSample", "BioProject", or "Both" field for a set of 
input SRA 'run_id'
}
\examples{
\donttest{
# SRA Library of interest
con <- SerratusConnect()
sra.df   <- get.sraMeta( 'SRR9968562' , con)
}
}
\keyword{Serratus}
\keyword{bioproject}
\keyword{biosample}
\keyword{palmid}
\keyword{sra}
