% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sraProj.R
\name{get.sraProj}
\alias{get.sraProj}
\title{get.sraProj}
\usage{
get.sraProj(run_ids, con, exclude.input.runs = FALSE)
}
\arguments{
\item{run_ids}{character, SRA 'run_id'}

\item{con}{pq-connection, use SerratusConnect()}

\item{exclude.input.runs}{boolean, exclude runs input 'run_ids'}
}
\value{
data.frame, run_id, biosample character vectors
}
\description{
Retrieve the "BioProject" field for a set of SRA 'run_id'
and return all SRA run_id for each unique BioProject
<UNDER DEVELOPMENT>
}
\examples{
\donttest{
# SRA Library of interest
con <- SerratusConnect()
library.bioProject   <- get.sraProj( 'SRR9968562' , con)
}
}
\keyword{BioProject}
\keyword{SRA}
\keyword{Serratus}
\keyword{palmid}
