% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.taxRunlist.R
\name{get.taxRunlist}
\alias{get.taxRunlist}
\title{get.taxRunlist}
\usage{
get.taxRunlist(genus = NULL, con = SerratusConnect())
}
\arguments{
\item{genus}{character, exact character sequence}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
data.frame, run_ids
}
\description{
Given an input taxonomic term, retrive all SRA RunInfo runs matching that
term
}
\examples{
\donttest{
# SRA Library of interest
con <- SerratusConnect()
homo.list   <- get.taxRunlist( genus = 'Homo', con)
}
}
\keyword{Serratus}
\keyword{palmid}
\keyword{sra}
\keyword{taxonomy}
