% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.palm.R
\name{graph.palm}
\alias{graph.palm}
\title{Convert virome.df into a palmprint-palmprint igraph object}
\usage{
graph.palm(
  sotu.vec = NA,
  pid.threshold = 30,
  expanded.graph = FALSE,
  all.edges = FALSE,
  get.metadata = TRUE,
  con = SerratusConnect()
)
}
\arguments{
\item{sotu.vec}{character, vector of sOTU id to retrieve}

\item{pid.threshold}{numeric, min percent identity to include an edge [30]}

\item{expanded.graph}{logical, return non-input sOTU joined to >=2 input sOTU}

\item{all.edges}{logical, return all sOTU with alignment to input sOTU}

\item{get.metadata}{logical, return sOTU meta-data for all returned entries}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
graph.virome, an igraph object
}
\description{
Convert virome.df into a palmprint-palmprint igraph object
}
\examples{

\donttest{
graph.palm(virome.df$sotu)
}

}
\keyword{Serratus}
\keyword{bioproject}
\keyword{biosample}
\keyword{igraph}
\keyword{palmid}
\keyword{sql}
\keyword{sra}
