% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.virome.R
\name{graph.virome}
\alias{graph.virome}
\title{Convert virome.df into an igraph object}
\usage{
graph.virome(
  virome.df = NA,
  nodes = "bio_project",
  edges = "tax_family",
  edge.threshold = 1,
  node.col = "bio_project",
  edge.col = "sotu"
)
}
\arguments{
\item{virome.df}{data.frame, produced by the get.virome() function}

\item{nodes}{string, column in virome.df to use as nodes ["bio_project"]}

\item{edges}{string, column in virome.df to use as edges ["tax_family"]}

\item{edge.threshold}{numeric, minium number of edges required to report [1]}

\item{node.col}{string, column in virome.df to use for node colors ["bio_project"]}

\item{edge.col}{string, column in virome.df to use for edge colors ["sotu"]}
}
\value{
graph.virome, an igraph object
}
\description{
Convert virome.df into an igraph object
}
\examples{

\donttest{
graph.virome(virome.df, nodes = "run", edges = "sotu", node.col = "bio_project")
}

}
\keyword{Serratus}
\keyword{bioproject}
\keyword{biosample}
\keyword{igraph}
\keyword{palmid}
\keyword{sql}
\keyword{sra}
