% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.virome2.R
\name{graph.virome2}
\alias{graph.virome2}
\title{Convert virome.df into an igraph object - all node}
\usage{
graph.virome2(virome.df = NA, node1 = "run", node2 = "sotu")
}
\arguments{
\item{virome.df}{data.frame, produced by the get.virome() function}

\item{node1}{string, column in virome.df to use as nodes ["bio_project"]}

\item{node2}{string, column in virome.df to use as edges ["tax_family"]}
}
\value{
graph.virome, a bipartite igraph object
}
\description{
Convert virome.df into an igraph object - all node
}
\examples{

\donttest{
graph.virome(virome.df, nodes = "run", edges = "sotu", node.col = "bio_project")
}

}
\keyword{Serratus}
\keyword{bioproject}
\keyword{biosample}
\keyword{igraph}
\keyword{palmid}
\keyword{sql}
\keyword{sra}
