% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.top10.R
\name{makeTop10}
\alias{makeTop10}
\title{make.top10}
\usage{
makeTop10(invec, top.n = 10, rename = "Other")
}
\arguments{
\item{invec}{factor, unordered and repeated entries}

\item{top.n}{numeric, how many entries to return before grouping to "Other" [10]}

\item{rename}{character, the string to return instead of ["Other"]}
}
\value{
factor, of the same length as invec, but with n+1 entries renamed "Other"
}
\description{
Aggregate a factor/vector and return only the top N entries
with all other entries renamed "Other"
Returned levels are rank-ordered by abundance for plotting
}
\examples{

## R Code Example

testvec <- c("A", "A", "A", "B", "C", "C", "C", "C")

makeTop10( testvec, top.n = 2 )

}
\keyword{Serratus}
\keyword{palmdb}
\keyword{palmid}
\keyword{sOTU}
