% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.palm.R
\name{graph.palm}
\alias{graph.palm}
\title{Convert virome.df into a palmprint-palmprint igraph object}
\usage{
graph.palm(
  virome.df = NA,
  pid.threshold = 30,
  expanded.graph = FALSE,
  all.edges = FALSE,
  con = SerratusConnect()
)
}
\arguments{
\item{virome.df}{data.frame, produced by the get.virome() function}

\item{pid.threshold}{numeric, min percent identity to include an edge [40]}

\item{con}{pq-connection, use SerratusConnect()}
}
\value{
graph.virome, an igraph object
}
\description{
Convert virome.df into a palmprint-palmprint igraph object
}
\examples{

\donttest{
graph.virome(virome.df, nodes = "run", edges = "sotu", node.col = "bio_project")
}

}
\keyword{Serratus}
\keyword{bioproject}
\keyword{biosample}
\keyword{igraph}
\keyword{palmid}
\keyword{sql}
\keyword{sra}
