% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.virome2.R
\name{graph.virome2}
\alias{graph.virome2}
\title{Convert virome.df into an igraph object - all node}
\usage{
graph.virome2(
  virome.df = NA,
  node1 = "run",
  node2 = "sotu",
  node1.col = "bio_project",
  node2.col = "tax_family",
  edge.col = "node_coverage"
)
}
\arguments{
\item{virome.df}{data.frame, produced by the get.virome() function}

\item{node1}{string, column in virome.df to use as nodes ["bio_project"]}

\item{node2}{string, column in virome.df to use as edges ["tax_family"]}

\item{edge.col}{string, column in virome.df to use for edge colors ["sotu"]}

\item{edge.threshold}{numeric, minium number of edges required to report [1]}

\item{node.col}{string, column in virome.df to use for node colors ["bio_project"]}
}
\value{
graph.virome, an igraph object
}
\description{
Convert virome.df into an igraph object - all node
}
\examples{

\donttest{
graph.virome(virome.df, nodes = "run", edges = "sotu", node.col = "bio_project")
}

}
\keyword{Serratus}
\keyword{bioproject}
\keyword{biosample}
\keyword{igraph}
\keyword{palmid}
\keyword{sql}
\keyword{sra}
