% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.virome2.R
\name{graph.virome2}
\alias{graph.virome2}
\title{Convert virome.df into an Virome igraph object - all nodes}
\usage{
graph.virome2(virome.df = NA, no.stats = FALSE, con = SerratusConnect())
}
\arguments{
\item{virome.df}{data.frame, produced by the get.virome() function}

\item{no.stats}{logical, do not calculate statistical tests [FALSE]}

\item{con}{pq-connection, use SerratusConnect()}

\item{node1}{string, column in virome.df to use as nodes ["run"]}

\item{node2}{string, column in virome.df to use as edges ["sotu"]}
}
\value{
graph.virome, a bipartite igraph object
}
\description{
Convert virome.df into an Virome igraph object - all nodes
}
\examples{

\donttest{
graph.virome(virome.df, nodes = "run", edges = "sotu", node.col = "bio_project")
}

}
\keyword{Serratus}
\keyword{bioproject}
\keyword{biosample}
\keyword{igraph}
\keyword{palmid}
\keyword{sql}
\keyword{sra}
